test(onboard): split openshell version coverage
